rederict the router, tidy up the nav bar, remove image folder under public and rename components
